refactor[#345] : 이메일 회원가입 api 요청 함수에 apiClient 적용
