+++ src/features/auth/api/index.ts
@@ -4,3 +4,4 @@ export * from "./postAddPetInfo.ts";
 export * from "./putChangeNickname.ts";
 export * from "./postSendCode.ts";
 export * from "./postCheckCode.ts";
+export * from "./postSignUpByEmail.ts";
