+++ src/features/auth/constants/endpoint.ts
@@ -7,7 +7,7 @@ export const LOGIN_END_POINT = {
 };
 
 export const SIGN_UP_END_POINT = {
-  EMAIL: `${API_BASE_URL}/users`,
+  EMAIL: "/users",
   VERIFICATION_CODE: "/users/auth",
   CHECK_VERIFICATION_CODE: "/users/auth/check",
   DUPLICATE_NICKNAME: `${API_BASE_URL}/users/nickname`,
